Attempt to fix #305 "undefined method `id'"-Error
I use better_errors and subsequent requests to the better_errors
middleware, e.g. trying to access the basic REPL depend on the instance
variable `@error_page` to be present in the BetterErrors::Middleware
instance.

Though as far as I understand the Rack- and BetterErrors-Code the
@error_page object is not "persisted" for inter-request access and right
now I dont understand how this usually works e.g. in a rails app
environment.

Nevertheless. My attempt to fix this is introducing a class variable
`latest_error_page` that serves as a temporary inter-request storage.
The current test suite does not brake and it fixes my problem.

I'd love to have some feedback about the inner workings of better_errors
and if @charliesome finds the time a little explanation how the
`@error_page` object is usually persisted for inter-request access.